[exec-mgr] Improve directory structure
