+++ server/executor_manager/manager.cpp
@@ -15,295 +15,20 @@
 #include "manager.hpp"
 #include "rdmalib/rdmalib.hpp"
 
-namespace executor {
+namespace rfaas::executor_manager {
 
-  Client::Client(std::unique_ptr<rdmalib::Connection> conn, ibv_pd* pd): //, Accounting & _acc):
-    connection(std::move(conn)),
-    allocation_requests(RECV_BUF_SIZE),
-    rcv_buffer(RECV_BUF_SIZE),
-    accounting(1),
-    //accounting(_acc),
-    allocation_time(0),
-    _active(false)
-  {
-    // Make the buffer accessible to clients
-    memset(accounting.data(), 0, accounting.data_size());
-    accounting.register_memory(pd, IBV_ACCESS_LOCAL_WRITE | IBV_ACCESS_REMOTE_WRITE | IBV_ACCESS_REMOTE_ATOMIC);
-    // Make the buffer accessible to clients
-    allocation_requests.register_memory(pd, IBV_ACCESS_LOCAL_WRITE | IBV_ACCESS_REMOTE_WRITE);
-    // Initialize batch receive WCs
-    connection->initialize_batched_recv(allocation_requests, sizeof(rdmalib::AllocationRequest));
-    rcv_buffer.connect(connection.get());
-  }
-
-  //void Client::reinitialize(rdmalib::Connection* conn)
-  //{
-  //  connection = conn;
-  //  // Initialize batch receive WCs
-  //  connection->initialize_batched_recv(allocation_requests, sizeof(rdmalib::AllocationRequest));
-  //  rcv_buffer.connect(conn);
-  //}
-
-  void Client::reload_queue()
-  {
-    rcv_buffer.refill();
-  }
-
-  void Client::disable(int id)
-  {
-    rdma_disconnect(connection->_id);
-    // First, we check if the child is still alive
-    int status;
-    auto b= std::chrono::high_resolution_clock::now();
-    kill(executor->id(), SIGKILL);
-    waitpid(executor->id(), &status, WUNTRACED);
-    auto e= std::chrono::high_resolution_clock::now();
-    spdlog::info("Waited for child {} ms", std::chrono::duration_cast<std::chrono::milliseconds>(e-b).count());
-
-    //auto status = executor->check();
-    //spdlog::info("executor {} {}", executor->id(), std::get<0>(status) == ActiveExecutor::Status::RUNNING);
-    //if(std::get<0>(status) != ActiveExecutor::Status::RUNNING) {
-    //  if(std::get<0>(status) != ActiveExecutor::Status::FINISHED)
-    //    spdlog::info(
-    //      "Executor at client {} exited, status {}",
-    //      id, std::get<1>(status)
-    //    );
-    //  else
-    //    spdlog::error(
-    //      "Executor at client {} failed, status {}",
-    //      id, std::get<1>(status)
-    //    );
-    //} else {
-    //  // FIXME: kill executor
-
-    //}
-    executor.reset();
-    spdlog::info(
-      "Client {} exited, time allocated {} us, polling {} us, execution {} us",
-      id, allocation_time,
-      accounting.data()[0].hot_polling_time,
-      accounting.data()[0].execution_time
-    );
-    //acc.hot_polling_time = acc.execution_time = 0;
-    // SEGFAULT?
-    //ibv_dereg_mr(allocation_requests._mr);
-    connection->close();
-    connection = nullptr;
-    _active=false;
-  }
-
-  bool Client::active()
-  {
-    // Compiler complains for some reason
-    //return connection.operator bool();
-    return _active;
-  }
-  
-  ActiveExecutor::~ActiveExecutor()
-  {
-    delete[] connections; 
-  }
-
-  ProcessExecutor::ProcessExecutor(int cores, ProcessExecutor::time_t alloc_begin, pid_t pid):
-    ActiveExecutor(cores),
-    _pid(pid)
-  {
-    _allocation_begin = alloc_begin;
-    // FIXME: remove after connection
-    _allocation_finished = _allocation_begin;
-  }
-
-  std::tuple<ProcessExecutor::Status,int> ProcessExecutor::check() const
-  {
-    int status;
-    pid_t return_pid = waitpid(_pid, &status, WNOHANG);
-    if(!return_pid) {
-      return std::make_tuple(Status::RUNNING, 0);
-    } else {
-      if(WIFEXITED(status)) {
-        return std::make_tuple(Status::FINISHED, WEXITSTATUS(status));
-      } else if (WIFSIGNALED(status)) {
-        return std::make_tuple(Status::FINISHED_FAIL, WTERMSIG(status));
-      } else {
-        // Unknown problem
-        return std::make_tuple(Status::FINISHED_FAIL, -1);
-      }
-    }
-  }
-
-  int ProcessExecutor::id() const
-  {
-    return static_cast<int>(_pid);
-  }
-
-  ProcessExecutor* ProcessExecutor::spawn(
-    const rdmalib::AllocationRequest & request,
-    const ExecutorSettings & exec,
-    const executor::ManagerConnection & conn
-  )
-  {
-    static int counter = 0;
-    auto begin = std::chrono::high_resolution_clock::now();
-    //spdlog::info("Child fork begins work on PID {} req {}", mypid, fmt::ptr(&request));
-    std::string client_addr{request.listen_address};
-    std::string client_port = std::to_string(request.listen_port);
-    //spdlog::error("Child fork begins work on PID {} req {}", mypid, fmt::ptr(&request));
-    std::string client_in_size = std::to_string(request.input_buf_size);
-    std::string client_func_size = std::to_string(request.func_buf_size);
-    std::string client_cores = std::to_string(request.cores);
-    std::string client_timeout = std::to_string(request.hot_timeout);
-    //spdlog::error("Child fork begins work on PID {}", mypid);
-    std::string executor_repetitions = std::to_string(exec.repetitions);
-    std::string executor_warmups = std::to_string(exec.warmup_iters);
-    std::string executor_recv_buf = std::to_string(exec.recv_buffer_size);
-    std::string executor_max_inline = std::to_string(exec.max_inline_data);
-    std::string executor_pin_threads;
-    if(exec.pin_threads >= 0)
-      executor_pin_threads = std::to_string(0);//counter++);
-    else
-      executor_pin_threads = std::to_string(exec.pin_threads);
-    bool use_docker = exec.use_docker;
-
-    std::string mgr_port = std::to_string(conn.port);
-    std::string mgr_secret = std::to_string(conn.secret);
-    std::string mgr_buf_addr = std::to_string(conn.r_addr);
-    std::string mgr_buf_rkey = std::to_string(conn.r_key);
-
-    int mypid = fork();
-    if(mypid < 0) {
-      spdlog::error("Fork failed! {}", mypid);
-    }
-    if(mypid == 0) {
-      mypid = getpid();
-      auto out_file = ("executor_" + std::to_string(mypid));
-
-      spdlog::info("Child fork begins work on PID {}, using Docker? {}", mypid, use_docker);
-      int fd = open(out_file.c_str(), O_RDWR | O_CREAT, S_IRUSR | S_IWUSR);
-      dup2(fd, 1);
-      dup2(fd, 2);
-      if(!use_docker) {
-        const char * argv[] = {
-          "executor",
-          "-a", client_addr.c_str(),
-          "-p", client_port.c_str(),
-          "--polling-mgr", "thread",
-          "-r", executor_repetitions.c_str(),
-          "-x", executor_recv_buf.c_str(),
-          "-s", client_in_size.c_str(),
-          "--pin-threads", executor_pin_threads.c_str(),
-          "--fast", client_cores.c_str(),
-          "--warmup-iters", executor_warmups.c_str(),
-          "--max-inline-data", executor_max_inline.c_str(),
-          "--func-size", client_func_size.c_str(),
-          "--timeout", client_timeout.c_str(),
-          "--mgr-address", conn.addr.c_str(),
-          "--mgr-port", mgr_port.c_str(),
-          "--mgr-secret", mgr_secret.c_str(),
-          "--mgr-buf-addr", mgr_buf_addr.c_str(),
-          "--mgr-buf-rkey", mgr_buf_rkey.c_str(),
-          nullptr
-        };
-        int ret = execvp(argv[0], const_cast<char**>(&argv[0]));
-        if(ret == -1) {
-          spdlog::error("Executor process failed {}, reason {}", errno, strerror(errno));
-          close(fd);
-          exit(1);
-        }
-      } else {
-        //const char * argv[] = {
-        //  "docker_rdma_sriov", "run",
-        //  "--rm",
-        //  "--net=mynet", "-i", //"-it",
-        //  // FIXME: make configurable
-        //  "--ip=148.187.105.220",
-        //  // FIXME: make configurable
-        //  "--volume", "/users/mcopik/projects/rdma/repo/build_repo2:/opt",
-        //  // FIXME: make configurable
-        //  "rdma-test",
-        //  "/opt/bin/executor",
-        //  "-a", client_addr.c_str(),
-        //  "-p", client_port.c_str(),
-        //  "--polling-mgr", "thread",
-        //  "-r", executor_repetitions.c_str(),
-        //  "-x", executor_recv_buf.c_str(),
-        //  "-s", client_in_size.c_str(),
-        //  "--pin-threads", "true",
-        //  "--fast", client_cores.c_str(),
-        //  "--warmup-iters", executor_warmups.c_str(),
-        //  "--max-inline-data", executor_max_inline.c_str(),
-        //  "--func-size", client_func_size.c_str(),
-        //  "--timeout", client_timeout.c_str(),
-        //  "--mgr-address", conn.addr.c_str(),
-        //  "--mgr-port", mgr_port.c_str(),
-        //  "--mgr-secret", mgr_secret.c_str(),
-        //  "--mgr-buf-addr", mgr_buf_addr.c_str(),
-        //  "--mgr-buf-rkey", mgr_buf_rkey.c_str(),
-        //  nullptr
-        //};
-        const char * argv[] = {
-          "docker_rdma_sriov", "run",
-          "--rm",
-          "--net=mynet", "-i", //"-it",
-          // FIXME: make configurable
-          "--ip=148.187.105.250",
-          // FIXME: make configurable
-          "--volume", "/users/mcopik/projects/rdma/repo/build_repo2:/opt",
-          // FIXME: make configurable
-          "rdma-test",
-          "/opt/bin/executor",
-          "-a", client_addr.c_str(),
-          "-p", client_port.c_str(),
-          "--polling-mgr", "thread",
-          "-r", executor_repetitions.c_str(),
-          "-x", executor_recv_buf.c_str(),
-          "-s", client_in_size.c_str(),
-          "--pin-threads", executor_pin_threads.c_str(),
-          "--fast", client_cores.c_str(),
-          "--warmup-iters", executor_warmups.c_str(),
-          "--max-inline-data", executor_max_inline.c_str(),
-          "--func-size", client_func_size.c_str(),
-          "--timeout", client_timeout.c_str(),
-          "--mgr-address", conn.addr.c_str(),
-          "--mgr-port", mgr_port.c_str(),
-          "--mgr-secret", mgr_secret.c_str(),
-          "--mgr-buf-addr", mgr_buf_addr.c_str(),
-          "--mgr-buf-rkey", mgr_buf_rkey.c_str(),
-          nullptr
-        };
-        int ret = execvp(argv[0], const_cast<char**>(&argv[0]));
-        if(ret == -1) {
-          spdlog::error("Executor process failed {}, reason {}", errno, strerror(errno));
-          close(fd);
-          exit(1);
-        }
-
-      }
-      //close(fd);
-      exit(0);
-    }
-    if(counter == 36)
-      counter = 0;
-    return new ProcessExecutor{request.cores, begin, mypid};
-  }
-
-  Manager::Manager(std::string addr, int port, std::string server_file, const ExecutorSettings & settings):
+  Manager::Manager(Settings & settings):
     //_clients_active(0),
     _q1(100), _q2(100),
-    _state(addr, port, 32, true),
-    _status(addr, port),
-    _settings(settings),
-    //_accounting_data(MAX_CLIENTS_ACTIVE),
-    _address(addr),
-    _port(port),
+    _state(settings.device->ip_address, settings.rdma_device_port,
+        settings.device->default_receive_buffer_size, true),
+    _settings(settings.exec),
+    _address(settings.device->ip_address),
+    _port(settings.rdma_device_port),
     // FIXME: randomly generated
     _secret(0x1234),
     _ids(0)
-  {
-    //_clients.reserve(this->MAX_CLIENTS_ACTIVE);
-    std::ofstream out(server_file);
-    _status.serialize(out);
-
-  }
+  {}
 
   void Manager::start()
   {
