+++ server/executor_manager/opts.cpp
@@ -1,0 +1,27 @@
+
+#include <cxxopts.hpp>
+
+#include "manager.hpp"
+
+namespace rfaas::executor_manager {
+
+  Options opts(int argc, char ** argv)
+  {
+    cxxopts::Options options("rfaas-executor-manager", "Handle client connections and allocation of executors.");
+    options.add_options()
+      ("c,config", "JSON input config.",  cxxopts::value<std::string>())
+      ("device-database", "JSON configuration of devices.", cxxopts::value<std::string>())
+      ("v,verbose", "Verbose output", cxxopts::value<bool>()->default_value("false"))
+    ;
+    auto parsed_options = options.parse(argc, argv);
+
+    Options result;
+    result.json_config = parsed_options["config"].as<std::string>();
+    result.device_database = parsed_options["device-database"].as<std::string>();
+    result.verbose = parsed_options["verbose"].as<bool>();
+
+    return result;
+  }
+
+
+}
