+++ server/opts.cpp
@@ -2,43 +2,6 @@
 #include <cxxopts.hpp>
 
 #include "server.hpp"
-#include "manager.hpp"
-
-namespace executor {
-
-  Options opts(int argc, char ** argv)
-  {
-    cxxopts::Options options("rfaas-executor-manager", "Handle client connections and allocation of executors.");
-    options.add_options()
-      ("a,address", "Use selected address.", cxxopts::value<std::string>())
-      ("p,port", "Use selected port.", cxxopts::value<int>()->default_value("0"))
-      ("pin-threads", "Pin worker threads to CPU cores", cxxopts::value<int>()->default_value("-1"))
-      ("use-docker", "Allocate executors in a Docker container.", cxxopts::value<bool>()->default_value("false"))
-      ("f,file", "Output server status.", cxxopts::value<std::string>())
-      ("v,verbose", "Verbose output", cxxopts::value<bool>()->default_value("false"))
-      ("r,repetitions", "Repetitions to execute", cxxopts::value<int>()->default_value("1"))
-      ("warmup-iters", "Number of warm-up iterations", cxxopts::value<int>()->default_value("1"))
-      ("max-inline-data", "Maximum size of inlined message", cxxopts::value<int>()->default_value("0"))
-      ("x,requests", "Size of recv buffer", cxxopts::value<int>()->default_value("32"))
-    ;
-    auto parsed_options = options.parse(argc, argv);
-
-    Options result;
-    result.address = parsed_options["address"].as<std::string>();
-    result.port = parsed_options["port"].as<int>();
-    result.server_file = parsed_options["file"].as<std::string>();
-    result.verbose = parsed_options["verbose"].as<bool>();
-    result.exec.pin_threads = parsed_options["pin-threads"].as<int>();
-    result.exec.use_docker = parsed_options["use-docker"].as<bool>();
-    result.exec.repetitions = parsed_options["repetitions"].as<int>();
-    result.exec.warmup_iters = parsed_options["warmup-iters"].as<int>();
-    result.exec.recv_buffer_size = parsed_options["requests"].as<int>();
-    result.exec.max_inline_data = parsed_options["max-inline-data"].as<int>();
-
-    return result;
-  }
-
-}
 
 namespace server {
